Dedupe feed items. Add playlist artwork.
Feed endpoint ignores the client limit for now,
hardcoded to 40 per request.

Currently time span is hardcoded to last 30 days.
Client can either specify a time cursor,
or we could try to expand the time range based on offset param.

There are a few tricky details when expanding the time range
with the de-dupe logic...
Will have to think about that a bit.