[API-69] Implement managed users endpoint (#69)
Follow on to #65 which implements the managed users endpoint.

The query is very similar but it needs to select the user by id and join
grants by their wallet address. And the resulting JSON for each record
uses `{ user, grant }` instead of `{ manager, grant }`.

Fun thing: My original queries were missing parens around the OR clause
at the end and that duped me into thinking things were really broken...

I also learned that if you put the `IS NULL` check first on a
`sqlc.narg()` param, it won't generate the correct type (will give you
an `interface {}` instead of a nullable boolean). But if it's last,
things work correctly 🤷