[API-116] Migrate underground trending endpoint (#73)
The python implementation does the score calculation on every request
instead of using indexed scores from the task. It's quite intense. This
version leverages the existing trending index but filters the results to
present lesser-known users.

There is a chance for duplicates across the two lists with this method,
but that's a reasonable compromise for now to get a much faster endpoint
without trying to port over all the complicated in-memory scoring logic.