[API-117] Add trending playlists endpoint (#76)
CANNOT BE MERGED until
https://github.com/AudiusProject/audius-protocol/pull/12111 is deployed
to production.

Adds the trending playlists endpoint, which is a pretty simple join of
the trending playlists scores against playlists that are public/not
albums/not deleted/etc.

I manually modified the schema file to include the new table, but we
could / should probably do a new pg_dump for it at some point.