Attempt at bulk fetching track access (#79)
This adds a bulk access fetch function and updates the FullKeyed
implementation to use it.

I opted for now to just fetch purchased playlists without trying to
filter down by tracks that have already been purchased. I'm open to
trying to make this a little smarter and skip that step if needed. But
it doesn't seem to significantly increase the query time for now. I
think there would be more work to do to make the purchased playlists
queries more efficient anyway.

Update: Per @stereosteve 's suggestion, changed this to use an error
group to do the queries in parallel.
Also, updated usdc-purchases query to use a materialized CTE for
usdc-gated tracks, since it's a very very small percentage of the number
of table rows and it _significantly_ speeds up the query for that
endpoint.