v1ChallengesUndisbursed (#80)
2-10x faster than the python because I don't sort by
completed_blocknumber first, but user id and challenge id. IIRC the
completed blocknumber sorting made sense for when we had identity doing
the claims on users' behalf, but we don't have that anymore so rather
than making an index there I just moved it to the last part of the sort.