remove unnecessary copies (#98)
Howdy it's me, ya boi

Given some of these methods had the signature `func (api ApiServer)`,
this could result in a full copy of the `ApiServer` struct. This means
that since the otter caches on the inside were also not pointers they
could be copied as well. Resulting in unwanted cache behavior. I
adjusted these routes to use pointers to ApiServer which I think
would've fixed most of the potential issues. I also made the otter
caches pointers and other relevant structs so as to avoid unnecessary
copies.

LMK if this copying was intentional. I was browsing this codebase to go
use otter in [audiusd](https://github.com/AudiusProject/audiusd) and
stumbled upon this.