[API-102] Transactions endpoints (#96)
_The infamous **thrice-deferred** PR, mythic and legendary_

And you'd think for all of that it would be really cool. But it's just
some endpoints that have a lot of finicky sorting capabilities (which
are _different between the two for some reason_).

And then I also had to zhuzh up the test framework a bit to support
testing all the sorting correctly.

Anyway, here's "Wonderwall".