[API-176] Adds validation and default value support for query params (#99)
Pulled in validator and defaults packages to allow us to do
struct-driven defaults and validation.
There are so many ways to expose this and I was aiming for it to be easy
in the places we have to repeat it a bunch:
* Allow endpoint code to define a struct with the query params it
supports, their default values, and their validation parameters
* Populating and validating said parameters is a single call on the
application. It will return a pre-formatted 400 error with the field
errors if things go wrong. Endpoint just returns the error up and lets
the error middleware handle it

Additionally:
* For now, I'm only adding a custom message for required fields.
Otherwise we just say the field is "invalid". We have to individually
implement messages for each validation tag if we want custom messages.
Easy to extend later if we want.
* Added a struct name lookup function that will use the 'query' tag for
the field name in the error message. Since the pattern for parsing query
params uses that tag to define the query param to read off the URL, this
keeps the output message in sync with whatever param name we use there.
* Funny quirk of the defaults package: For number fields with a default,
if you pass '0' in the query string, it will be interpreted as falsey
and use the default value instead. This seems nice to me :shrug:

I will acknowledge that this kind of binds us to the pattern of using a
struct to define query param parsing. I don't think we have any wild use
cases where this will be super annoying. And I like the consistency of
doing it this way.
If we have an endpoint that really really needs to do crazy stuff, you
can always parse query params individually, dump them into a validation
struct, and call the validator directly..


I only added the validation to one endpoint for this PR so I can get
feedback on the pattern first. Will follow up and implement
defaults/validation for at least `limit` in a separate PR.