[API-176] Migrate all query param parsing that needs validation to new pattern (#100)
Building off #99 , this migrates all of the endpoints that need
validation for query params to use the new pattern.
* Most limits have a max of 100, defaults vary by endpoint
* Most offsets have a max of 500, but used larger values for endpoints
where it makes sense to be able to get a lot more items (track library,
transaction history, etc). Open to suggestions on what feels appropriate
for these!
* For most of the sort methods, directions, filters, etc I looked into
the python code base to find the values that were allowed there and
duplicated them here. And at the very least, we are matching the values
that the SQL query building switches over. Happy to yank out this
validation if it feels too sketchy and just let strings be strings.