[API-182] Fix user tracks endpoint params/sorting (#114)
Two issues here:
* The user tracks handler is wired up to the `/users/:userId/tracks`
route but it doesn't support querying by ID and returns nothing. Updated
it to handle either case based on presence of the `handle` param in the
route.
* The sorting logic isn't correct. This endpoint supports both the
`sort` (deprecated) and `sort_method` params and is supposed to prefer
the latter if specified. When using `sort`, the options are `plays` and
`date`, defaulting to `date`. So I added back the aggregate plays join
to support this case.

TODO:
* Consider conditionally joining aggregate plays only if sorting by it
* TESTS. As those would have caught this :-)