Remove global test app... isolated test db per test. (#117)
Currently this takes test runtime from like 2s to like 6s... which is
sad.

But I want to merge it because it is a big diff that is likely to cause
conflicts.
It would be good to get everyone using the test app helpers... and I
have idea on how to make faster tomorrow.

After merging you have to:

```
docker compose down --volumes
docker compose up
```