Use base row table name convention. (#120)
Instead of defining many base rows at the top level with arbitrary
variable names... have a single top level `baseRows` variable which is a
map keyed by table name.

So when populating fixtures for a table, we key the base row using table
name by convention.

This also makes `FixtureMap` more flexible so you can create fixtures
for any table so long as it has a base row defined. This way we don't
need to do code changes to add fixtures for a given table (assuming we
have a base row)