Audiusd SDK (#123)
- Adds the audiusd sdk into the ApiServer struct, default variables
should suffice but in the future we'd want these to be rpc pools instead
of validators. I'll get those stabilized so we can make that change.
- Updated the imports to use the new versioned connect rpc types. Most
of these were found in the indexer anyway. We should find a way to keep
these up to date. Didn't realize y'all were importing the older versions
still.
- I hesitated on adding any kind of routines to this. Things to cache
node info, latest block, retry connections, etc. Happy to add that as a
follow on if there's a pattern somewhere in this repo already.