Fix cache pollution due to context parameter capturing (#129)
This was wild.

1. Fiber is [pretty clear in their
docs](https://docs.gofiber.io/#zero-allocation) about why you should not
hold on to any value in the Context past the life of the request.
2. We have a couple of places where we call `c.Param()` to get a string
value of a wallet or handle. We pass that into a helper function to do
some kind of lookup of a user.
3. To help with performance, we use Otter caches in those places, and
the usage is pretty straightforward (try to get value, if we miss look
it up and insert it into the cache before returning)

Imagine our surprise when the cache started returning the _wrong_ user,
sometimes but not all the time, and with no discernible pattern.

The _very_ subtle issue is that since we pass the value all the way down
the stack and _into_ the `Set` function on the cache without ever
assigning it to a new variable or making a copy, we end up inserting a
memory address belonging to a reusable Context from a pool as the key.
So at any arbitrary point in the future, we might reuse that same memory
address as the key for a different value. And so the performance
optimizations of two different libraries collude to remove some hair
from our heads.

TL/DR you should **_always_** copy a value before using it as an Otter
cache key, even if your inexperience tells you there's no way that value
is passed by reference. Because it most definitely doesn't work the way
you think it does.