Fix trending playlists query to respect sorting criteria (#130)
Matching the logic from
https://github.com/AudiusProject/audius-protocol/pull/12302

This updates eligibility to require >= 5 unique owners and tracks for a
playlist to be considered for trending.
The query is slightly different from the one in DN because this one
powers the UI and will be called more frequently. Did some
explain/analyze iterations to get to a version to uses materialized CTEs
(and I added an index to playlist_trending_scores) so we can get the
trending list back in around 800ms. There's a linear to improve this
later but it's usable for now!

Also updated the tests to use isolated fixtures.