[API-188] Add user ai_attributed_tracks endpoint (#132)
Added the /ai_attributed_tracks endpoint, which returns not AI tracks
created by the user in route params, but AI tracks with them attributed
as the source for training. Otherwise all the filters are the same. So I
decided to reuse the user/tracks endpoint logic and give it an extra
param that we force to `true` for the AI case.

Also implemented the `filter_tracks` query param, which we were missing
for this endpoint. It appears to be used in client, likely to allow you
to view your own unlisted tracks.

And bonus fix for an offset param validation I mistakenly removed in a
previous PR 😛