v1UsersTop and v1UsersGenreTop (#138)
Note: Unlike Python, the genres here are case sensitive. This keeps the
index working and makes the query fast. We could probably make a list of
all allowed genres somewhere and marshal the parameter to that before
querying though.

Also, the genre is now required in the top/genres endpoint.

Opted against adding tests here, at least for now.