[API-167] Add Isaac's "for you" endpoint (#148)
This gives you tracks from your top genres which you haven't listened to
before which exceed a certain trending score (based on passed in
time_range param), randomized and limited by the query params.

Thanks @isaacsolo for the base query work on this. It was immensely
helpful.
I tweaked it a bit to reduce the number of joins and to get it
outputting just track_ids since our pattern in Bridge is to query
track_ids first and then run those into a shared query pattern to
resolve to tracks.

Note: I chose to implement this as a sub-route of `/users` since it's
specific to a given user. Open to changing this, but it seemed more in
line with our patterns for things like /account, etc.

Also might tweak the way the minimum scoring works for the trending
section.

TODO:
* Tests, will be a little tricky but I have a plan