Enable unescaping for params to fix fetching playlists with emojis in slug (#154)
We fail to look up playlists by permalink if the slug has an emoji in
it. That's because the default behavior for fiber is _not_ to unescape
string param values. Since we store the slugs as UTF-8 encoded, we need
this on or we have to manually decode things everywhere where it might
matter.