Add support for 'query' in library endpoints (#159)
This adds the missing support for the `query` query param in
library/tracks, library/album, library/playlists. It resolves to an
`ILIKE` with a lowercased version of the query against playlist/track
name or user.name field of the owner.

I limited the query length to 250 because unbounded is a bad idea and
250 seems like more than enough to specify whatever filter string you
want to search on.