[API-228] Migrate to OAPI3.0 (#179)
I ran our swagger 2.0 JSON files through a converter and then fixed all
of the spots where it accidentally dropped `oneOf` definitions. To test
it, I updated the SDK generator to skip all of the intermediate
conversion and turned on validation. The output from that produced no
diff to the generated SDK files, so I would say we are good for now!

Note also that we're switching to .yaml (more human readable), which
makes the files smaller. But that is an _intentional_ deviation from
hosting a `swagger.json` file (i.e. api.audius.co/v1/swagger.json will
not work or might redirect to old discovery routes until we turn that
off). Since I believe we are the only ones who use that output file, I'd
like to let this breaking change be fine.

Minor things:
* There are a few models that I think are technically not used and could
be removed. But my goal with this was to produce _no change_ to the
generated SDK output.
* We do still have some warnings in the generator about things like
`wildcard_split`. But again, it generates identical code, so leaving
this alone for now.