[API-171] Add feeling lucky endpoint (#183)
`/v1/tracks/feeling_lucky`

The original implementation in discovery just used a min_followers
filter that defaulted to 100. For the new Search/Explore page, we're
changing this to be an always-on filter that requires 250+ streams on a
track to help filter down the result set before we select a random
number of them. Local testing against prod db has this pretty reliably
under 500ms. There might be some optimizations that could bring this
down another 100ms or so, but it feels totally usable for now.

I chose to _not_ use the min_followers filter by default, as it actually
makes the query slower. But to preserve compatibility with the existing
endpoint, I implemented it as an optional query param that will add in a
JOIN and additional filter condition if you specify it.