Add db-backed HLL to track unique metrics (#202)
HyperLogLog is a cool data structure and we're stealing the impl from
our friends at Axiom.
With high accuracy, we can probabilistically store unique counts on IP
addresses very cheaply!

When reviewing this PR, to understand usage, the test
`TestHLL_MergeExistingSketch` is a great place to start!

Here's a great video if you need a reference :-)
https://www.youtube.com/watch?v=lJYufx0bfpw

Tested:
- Little integration tests w/ the bridge DB