Hash encode specifiers.  Support group / timeset pagination (#233)
Notification specifier can be either a string or a number. If it's a
number we should hash encode it.

So coming out of the db, the value is always a string, and if it is a
quoted int, HashifyJson will encode it.

Tested by hardcoding `userId = 1` in `v1_notifications.go` and
confirming notifs load in client.

---

group / timestamp logic ported from python mostly verbatim.