Solana Indexer: Caching (#247)
Because the processor gets fired for the same transaction if multiple
accounts are included from the tracked mints, it makes sense to cache
these transaction responses to not abuse our RPC budget. The cache has a
hit ratio of about 66%, which makes a bit of sense considering for every
balance change in one account there's at least one balance change in
another of the same mint.

Also fixes a bug where the mint filtering wasn't working, opting to
separate the concerns and caching the retrieved artist_coins until the
subscription blows away the cache.

Also:
- Fixes a bug where payment router wasn't ignoring conflicts.
- Removes batch sleep (tx sleep is enough)
- Fixes the "prod" test to use a non-JS interpreted response
- Adds a test to ensure the mint filters work