Add slot to deadletter queue (#296)
An oversight of the deadletter queue impl had it inserting its data at
slot 0. This would cause none of the updates to trigger for current
balances etc because they use slot to order not insertion time.

Adds slot to the `sol_unprocessed_txs` table and uses it in the retry