[API-135] Fix various notification issues (#297)
Commits can be reviewed separately if you'd like.

3 Changes:

1. Fix bug in seen_at grouping from previous implementation in
https://github.com/AudiusProject/api/pull/287. Currently with the
checked in code, we are not grouping unread messages. Add some comments
in the sql to make this more obvious to future readers

2. Always return all notification types (change the param from
validTypes to types in future to prevent regression). Protocol PR here:
https://github.com/AudiusProject/audius-protocol/pull/12669

3. Filter out notifications for deleted tracks / playlists (API-135)
from a few months ago, which we said we could patch in the new bridge
impl when that shipped.