Add proxied support for selecting random read replica (#312)
This adds a `DBPools` class/struct/whatever you call these in golang
that selects a random replica from a list of connection strings in
config and proxies read requests through to it. Tests are hard-coded to
just write to the first replica. For write workflows, we can update this
to take in a write config string as well and then implement write
operations by proxying to that instead.

Trying to minimize impact here and get this out reasonably quick to fix
our issues with database load.

Unit tests passed, and I ran apiDiff locally with two read replicas and
it did not explode.