Don't use "latest" for arch dependent builds for simplicity (#319)
In main, the builds were doing both "latest" and sha-based tags on each
arch image, then trying to combine all four together into the "latest"
and sha-based tag for the combined manifest.

This PR splits it out so that the arch dependent builds are just sha
based, and an extra metadata extraction step gets their tags instead of
sharing the one for the outputs in the merge phase.