[API-321] Tests for comms mutations (#338)
* Migrated all but the blast tests over from protocol. These are all
query/db level tests that run the validator/processor code and then
check the results in the db.
* Changed from sqlx to pgx patterns for querying
* Updated the chat*() functions that write to the DB to support using a
transaction or a pool
* Updated the rate limiter to be just a config instead of a struct we
query values from. This was over-engineered because we don't actually
change it at runtime
* Updated processor/validator to take rate limit config as an arg so
that we can change it for the tests. They are randomly flaky when trying
to execute 10+ queries in under a second.
* Added some basic tests for the endpoint processing to make sure we are
decoding RPCs correctly.

Handful of TODOs to take care of, but wanted to get it up for review.