[API-290][API-311] Stub out deprecated comms RPC endpoints (#343)
We aren't doing peering in the bridge implementation of comms, so these
aren't needed.
However, for the initial implementation, I think it's a good idea to
implement them to maintain backwards compatibility. They will return
empty responses and log warnings if called so that we can track down any
bad routing that happens.