[API-205][QA-2309] Limit trending playlists track fetching to 5 items (#344)
This restores a behavior we had on DN to limit the number of tracks
fetched in the `full` logic to just the first five.
This prevents really large playlists from slowing down the response
times. But it does mean playlist tiles are only going to play 5 items
from the playlist unless we otherwise fetch the remaining tracks on
client.
Chatted with @dylanjeffers about this and it's a reasonable trade off
for now.