[API-321] Migrate tests for blast writes (#342)
This is a followup to #338 to migrate the quite sizable blasts test
file. Source is here:
https://github.com/AudiusProject/audius-protocol/blob/ab73d979f98e151ac0aa408c7a82c105c3b6355e/comms/discovery/rpcz/chat_blast_test.go#L760

* Migrated everything to pgx patterns
* Split the file into multiple isolated tests instead of one huge test
* Using fixtures instead of manual inserts for setup data
* Removed a lot of the 'in the past' time variables in favor of just
using time.Now() for operations that are sequential
* Updated the version of `getNewBlasts` in the comms directory to match
the one we use in the get endpoint. We are being intentionally WET here.
* Fixed an issue with upgrading blasts to chats where the timestamp that
we assign to the new message would be in the wrong timezone if run on a
machine not set to UTC. In practice, this won't break on CI or
production, but it _does_ break when running on a local machine 😢

As a bonus, also cleaned up the tests from the previous PR to match the
patterns here.