[PE-6613] Add `/coins/ticker/:ticker` in order to enable asset pages to use ticker  (#362)
Currently, we use the `getCoin` endpoint for asset pages, but we go
based off of mint. Creating a new endpoint so we can use ticker and
avoid any convoluted FE logic to replace mint with ticker. This will
essentially allow us to navigate to audius.co/coins/$AUDIO. The rest of
the endpoints are ok with just being mint based.


https://audius-internal.slack.com/archives/C08JHQH11R8/p1757016195240729