Fix coin endpoint behavior with bad input (#366)
`/v1/users/:id/coins/:mint` expects a mint address, not a ticker. But if
called with invalid input it still attempts to fetch price data. This
will result in a 400 and we will skip adding it to the price cache. This
means our next request is going to attempt to fetch it again, and we
will use up more quota than we would like due to missing the cache on
every request.
This updates the endpoint to bail early if the mint value isn't a valid
address.


Bonus change: Added parsing of the returned error message so we can
expose that in out logs. Made it easier to track down why we were
getting back a 400.