[API-316] Migrate /challenges/:id/info (#408)
This is mostly a debug/monitoring endpoint (not used in client or
exposed in SDK).
Fun thing I learned while migrating this: The implementation in python
doesn't do the math right....

I moved the calculations to be purely in the query and return the
remaining pool amount as an integer. Since these values are specified in
AUDIO and not in WEI, we don't need bigint/numerics for that if we cast
the final result before returning it from the query.