[API-347] Add health check with known healthy content nodes (#412)
We can/should add more to this in the future (like a quick DB health
check instead of our monitor hitting a user lookup endpoint directly).

But for now, this is meant to cover the one case where we consume the
health_check in client code: Getting the list of healthy storage nodes.
I based this on the [healthy nodes task in
discovery](https://github.com/AudiusProject/audius-protocol/blob/64ff2d9c28904fe2a31c57a275faa6eccdcd97d9/packages/discovery-provider/src/utils/get_all_nodes.py#L107-L114),
which runs every 2 minutes, attempts to ping all the content nodes, and
then caches the result.

Notable differences:
* Our source list is hard-coded in config and not read from the contract
each time (should update this in the future)
* We will retry each node thrice to give them a chance to respond
* We aren't using redis for this. Each replica will have its own
in-memory state for it. In practice, this is probably fine. This list is
used only for upload to help find a host to accept the file. But it gets
replicated to the correct rendezvous hosts afterwards.