Remove unnecessary check for no mints (#418)
Not sure why we had this check in the first place, but it scans the
whole artist coins table on every request, presumably to make sure we
return an empty array if there are no entries in the table.

The query works just fine without it and will return empty in that case.
I added a test to confirm this and also added tests on the other coins
list endpoints to make sure those work as expected when there are no
coins in the table (which should only be for a short period of time :-)
)