Update schema and fix test-schema command to remove restrict directives (#445)
Cleaning up schema (it got gen'd from prod again) and adding the `sed`
workaround to remove the restrict directives until sqlc gets patched.

Also removing the "schema dump" section from the readme because it
encourages dumping from an arbitrary database and we want this to come
from the test database with the migrations from DDL applied.