[API-355] Add health check metrics for new core indexer (#446)
This mimics the health check strategy we used for block diffs before:
Pass a query param to indicate max allowed and we return a 500 if it is
exceeded. Also will return 500 if we failed to compute the block diff
and a max was provided.
Otherwise, this is intended to not interrupt the health check since
that's also used by various consumers to get the list of content nodes.

I'm open to making this thing cache the block diff and only compute it
once every few seconds if we think the RPS to this endpoint will add too
much load. It does make a RPC call to core and a database request, so
it's not exactly the lightest weight thing.