Fix missing coin balance after creation (#448)
Because the artist coin is missing stats after creation until the stats
job runs (on a 5min interval right now) we should use a left join so
that it still shows up when insights are missing. The insights are used
in the query for getting price, as is the pool, so we'll have to do some
extra work here I think to prevent the price from looking like 0?

Testing this locally was a bit of a beast... I think we just get this
out and test against prod launch of a coin to see if this solves the
optimism problem since none of this is live yet anyway.