Update dbc pools without needing RPC (#457)
The RPC doesn't always have the state change by the time the
subscription notifies the indexer, causing the DBC to not be found.

The subscription actually gives us the updated state, so use that
instead.

Still need to fetch the config, but that's going to be stable and around
much longer and won't change as frequently.