[PE-7180] Add all time stats to coins job/responses (#475)
Adds a call to the [all time
trades](https://docs.birdeye.so/reference/get-defi-v3-all-time-trades-single)
endpoint to get all time data and merges that into the coin stats table
along with the overview info.

Notable:
* Coin stats job runs this new request and the overview request in
parallel with an error group. Goal was to not double the length of time
it takes to update stats for a coin.
* Updated the overview fetcher to also attempt to extract better error
messages (missed this when I fixed the dbc pools function a while ago)
* Some bonus schema upgrades that weren't added in in previous PRs