[PE-7202] Add support for coin_flair_mint on users (#490)
Adds the column to track preferred coin flair mint and logic to use it
for computing the badge on user records
* `null` uses existing 'auto' logic
* Empty string indicates 'none' and we return no badge even if you have
or hold a coin
* A valid mint string will prefer that coin if you hold a balance.
Otherwise it falls back to the default logic.

Also adds the column to the response for the user to facilitate the edit
profile flow.


Note: Needs an indexing change in the `apps` repo to actually set these.
I've only tested via unit tests for now.