[API-359] Adds user score indexer (#488)
This is an attempt to do a less disruptive update to
`aggregate_user.score`. The legacy DN implementation ran an update query
that would compute all user scores at once and then update them in the
aggregate_user table.
This is bad for at least the following reasons:
* Locks aggregate user table for the length of the query (which is on
the order of 5-10 minutes to complete)
* Reads from the write db to compute the scores, so subject to blocking
by all sorts of other queries

`get_user_scores` was also doing some stuff per-user that made the query
a little inefficient.

New plan is thus:
* Run an indexer job that computes user scores in batches from the read
replica, with a modified query that is a little more efficient against
batches of a few thousand users. The read query sorts by
created_at,user_id and returns a cursor we use to start the next batch.
* Indexer fetches ids and scores, dedupes them (we have some duplicate
user records even with the grouping and `distinct` is really slow here),
then pushes the updated ids/scores into the write replica.
* Job finishes when fetched size is < batch size (hit the beginning).
It'll loop back around immediately.

Some additional features added after PR feedback:
* Added tables to hold precomputed values for distinct hours and tracks
played as well as a score features table that currently holds number of
fast challenges (these have been backfilled manually)
* I updated the triggers for plays and challenges to update the new
precomputed aggregate tables. We can also truncate and recompute these
whenever we want if we want to change the logic.
* Added an index to aggregates that will help with a hot path in our
score computation (index needs to include both followers and follow
count, since we use those in this query).

Testing on local machine with a prod replica as a data source it takes ~
3 minutes, but that's without the new index on aggregate_user.

This is a halfway step between our existing slow query and something
that can update scores on a streaming basis.