Some small CoreIndexer improvements (#501)
I missed log syncing for the aggregates job, moving the sync on ticks up
a level to cover both goroutines.
Also updated the core processing loop to be resilient to panics. We will
now sleep a bit and try again.

Open question: 10 seconds too long of a sleep? Should we sleep in all
error cases? I'm mostly worried about spewing a bunch of logs in cases
where we have a temporary error.