Fix logging for real this time (#502)
I was not aware that syncing a logger doesn't sync loggers spawned from
it....
Aggregates calculator has its own logger.
I know we don't love this pattern, but it's what we have for now!


Also I meant to make one of those CoreIndexer logs a debug, it is
realllllly chatty.