[API-360] Implement top playlists endpoint (#503)
We missed this endpoint in migration. It looks like
`/v1/full/playlists/top?mood=upbeat&limit=20&offset=0`

It's basically a hard-coded custom search query. This is being added to
preserve backwards compatibility.


Also fixed a couple of issues with search:
* genre/mood are case sensitive and our genres/moods are currently
indexed as Capitalized words. Since the in-the-wild usage of this
endpoint passes mood in a lowercase format (and because it's so
incredibly reasonable to expect that to work), I added some logic to the
search helpers where applicable to normalize these values. Should also
fix any misses on the general search endpoints.
* Subtle lil' miss on the original implementation: in the `q.Moods > 0`
conditional, we were looping over `q.Genres` to apply the score boosts
for tracks in the playlist. So it didn't have the desired effect of
boosting playlists who had more tracks with the desired moods.