[PE-7261] Fix search by genre with multiple words (#506)
I made a change for the top playlists endpoint to lowercase and then
capitalize the genre for searching. That's not really necessary for that
endpoint and broke our existing genre search.
This removes the genre mangling code (safe to do with mood because they
are all single word) and adds some tests to catch this regression if it
happens again.